Adicionando Controller de Historico

+++ HealthyGarden.Domain/Entities/Historic.cs
@@ -6,7 +6,7 @@ namespace HealthyGarden.Domain.Entities
         public int Id { get; set; }
         public int GardenId { get; set; }
         public DateTime IrrigationDate { get; set; }
-        public decimal Moisture { get; set; }
-        public decimal Temperature { get; set; }
+        public short Moisture { get; set; }
+        public short Temperature { get; set; }
     }
 }
